move dedicated classes to test cases
